program2/requirements.txt
streamlit
yt-dlp
pydub
imageio-ffmpeg
audioop-lts
mailjet-rest
python-dotenv

